+++ packages/pagination/src/__tests__/web/pagination-wrapper.web.test.js
@@ -1,4 +1,0 @@
-import tests from "../test-wrapper-helper";
-import pagination from "../../pagination-wrapper";
-
-describe("Pagination Wrapper on Web", tests(pagination));