+++ packages/pagination/src/__tests__/web/pagination.web.test.js
@@ -1,5 +1,0 @@
-import "jest-styled-components";
-import tests from "../test-helper";
-import Pagination from "../../pagination";
-
-describe("Pagination on Web", tests(Pagination));