+++ packages/styleguide/src/__tests__/android/styleguide.android.test.js
@@ -1,5 +1,0 @@
-import shared from "../shared";
-
-describe("Styleguide tests for android", () => {
-  shared();
-});